Don't add client, portal IDs to auth redirect URL query
We don't know what HubSpot will do in the future. They might start scrubbing the
URL (as suggested by the recent report of the OAuth redirect security hole).

Consequently, this requires that we separate the portal ID and client ID from
the Auth type. This does actually simplify the code, though it now requires the
library user to track (at least) the client ID separately for refreshing.